Add files via upload
This code implements a simple neural network from scrach with one hidden layer to solve the MNIST digit classification problem using Keras dataset. The code starts by loading and preparing the dataset, then defines activation functions and their derivatives, and creates a NeuralNetwork class.
The NeuralNetwork class includes methods for forward propagation, computing cross-entropy loss, backpropagation, prediction, and training. The network consists of an input layer, a hidden layer, and an output layer. The ReLU activation function is used for the hidden layer, and the softmax activation function is used for the output layer.
The code initializes the neural network with random weights and biases, trains it on the training data for 500 epochs, and evaluates the test accuracy. The learning rate is set to 0.01.
The code also includes comments for each section, making it easier to understand the flow and the purpose of the code. This code serves as a starting point for implementing more complex neural networks and solving various machine learning problems.